% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{Arithmetic}
\alias{Arithmetic}
\alias{+}
\title{Arithmetic}
\usage{
a + b

"+"(a, b)
}
\arguments{
\item{a, b}{numeric vectors or objects which can be coerced to such.}
}
\description{
Obviously arithmetic in trumpverse doesn't work the same as you'd expect in the real world.
}
\details{
The unary arithmetic operators are a exist at the core of the R language.

But in the trumpverse, language is but a flexible and fluid paradigm which can be
tailored and corrupted to ridiculous ends.

Therefore it seems reasonable to introduce alt.reality versions of these basic operator
 functions. Enjoy
}
\examples{
 1 + 1  # may or may not equal 1

}
