% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{Comparison}
\alias{Comparison}
\alias{==}
\title{Relational operators}
\usage{
"=="(e1, e2)
}
\description{
Compare objects, test equality etc.
}
\details{
1 == 1 ? Well, maybe...
}
\examples{
 1 == 1
 [2] FALSE

}
