% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confuse.R
\name{confuse}
\alias{confuse}
\title{confuse}
\usage{
confuse(data)
}
\arguments{
\item{data}{}
}
\value{
Object of the same class
}
\description{
Confuses the values of a data frame or tibble. Essentially randomises the values in each column.
}
\examples{
  # Luke Skywalker is a brown haired droid from Alderaan..?
  confuse(dplyr::starwars)

}
