% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmin.R
\name{Extremes}
\alias{Extremes}
\alias{max}
\alias{min}
\title{Maxima and Minima}
\usage{
max(..., na.rm = FALSE)

min(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric or character arguments.}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\description{
Returns one of more extreme values from the data, but not necessarily the actual maximum or minimum.
}
\details{
`max` and `min` return the maximum or minumum of _some_ of the values present in their aguments.

If na.rm is FALSE an NA value in any of the selected arguments will cause a value of NA to be
returned, otherwise NA values are ignored.
}
\examples{
  max(1:5)  # may or may not equal 5

  # Which is bigger, 1.8 million, or 600,000? ... it depends
  max(1800000,600000)

}
