% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\docType{data}
\name{logical}
\alias{logical}
\alias{T}
\alias{F}
\title{Logical vectors}
\format{An object of class \code{logical} of length 1.}
\usage{
TRUE
  FALSE
  T; F

F
}
\description{
In trumpverse true is false is true. Sometimes
}
\details{
TRUE is always true, as FALSE is false. But be careful, since F is true, and T is false.
Sad.
}
\examples{
 1 == 1
 [1] TRUE

 T
 [1] FALSE

 (1 == 1) == TRUE
 [2] TRUE

 (1 == 1) == T
 [2] FALSE

}
\keyword{datasets}
